Move error logic to activation card
